chore: pin dependencies and cleanup code
