+++ Cargo.toml
@@ -4,12 +4,16 @@ resolver = "2"
 
 [patch.crates-io.biscuit]
 git = "https://github.com/beltram/biscuit"
-branch = "feat/wasm"
+tag = "v0.6.0-pre.core-crypto-0.6.0"
 
 [patch.crates-io.jwt-simple]
-git = "https://github.com/beltram/rust-jwt-simple"
-branch = "wire"
+git = "https://github.com/wireapp/rust-jwt-simple"
+tag = "v0.11.3-pre.core-crypto-0.6.0"
 
 [patch.crates-io.rcgen]
 git = "https://github.com/wireapp/rcgen"
-tag = "v1.1.0-pre.core-crypto-0.6.0"
+tag = "v1.2.0-pre.core-crypto-0.6.0"
+
+[patch.crates-io.x509-parser]
+git = "https://github.com/wireapp/x509-parser"
+tag = "v1.0.2-pre.core-crypto-0.6.0"
