+++ acme/Cargo.toml
@@ -12,20 +12,16 @@ serde = "1.0"
 serde_json = "1.0"
 thiserror = "1.0"
 rusty-jwt-tools = { version = "0.1", path = "../jwt" }
-jwt-simple = "0.11.2"
-base64 = "0.21.0"
+jwt-simple = "0.11.3"
+base64 = "0.21"
 url = { version = "2.3", features = ["serde"] }
 rcgen = "=0.9.2"
 time = { version = "0.3", features = ["serde", "serde-well-known", "wasm-bindgen"] }
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-reqwest = { version = "0.11", features = ["json"], default_features = false, optional = true }
+reqwest = { version = "0.11", features = ["json", "__rustls"], default_features = false, optional = true }
 testcontainers = { version = "0.14", optional = true }
-
-[dependencies.x509-parser]
-git = "https://github.com/wireapp/x509-parser"
-tag = "v1.0.1-pre.core-crypto-0.5.0"
-features = ["verify", "validate"]
+x509-parser = { version = "=0.14.0", features = ["verify", "validate"] }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3"
