+++ cli/Cargo.toml
@@ -11,6 +11,6 @@ publish = false
 rusty-jwt-tools = { version = "0.1", path = "../jwt" }
 clap = { version = "4.0", features = ["derive"] }
 anyhow = "1.0"
-jwt-simple = "0.11.2"
+jwt-simple = "0.11.3"
 serde_json = "1.0"
 console = "0.15"
