+++ ffi/Cargo.toml
@@ -13,10 +13,3 @@ crate-type = ["cdylib", "staticlib"]
 
 [dependencies]
 rusty-jwt-tools = { version = "0.1", path = "../jwt" }
-cfg-if = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-[features]
-default = []
-haskell = []
