+++ jwt/Cargo.toml
@@ -16,14 +16,12 @@ thiserror = "1"
 serde = "1"
 base64 = "0.21.0"
 uuid = { version = "1.1", features = ["v4"] }
-hex = "0.4"
 p256 = "0.12"
 p384 = "0.12"
 sec1 = "0.3"
 url = { version = "2.3", features = ["serde"] }
 serde_json = "1.0"
 sha2 = "0.10"
-oauth2 = "4.2"
 chrono = "0.4"
 either = { version = "1.8", features = ["serde"] }
 time = { version = "0.3", features = ["serde", "serde-well-known", "wasm-bindgen"] }
@@ -31,7 +29,7 @@ json-patch = "0.3"
 biscuit = "0.6.0-beta1"
 rand = "0.8"
 rand_chacha = "0.3"
-jwt-simple = "0.11.2"
+jwt-simple = "0.11.3"
 zeroize = { version = "1.5", features = ["zeroize_derive"] }
 # Fixes a cyclic dependency on getrandom (see https://github.com/tkaitchuck/aHash/issues/95#issuecomment-874150078)
 indexmap = "=1.9.2"
